Rename the workers to match the topic
