+++ src/net/services/Prism.Picshare.AzureServices.Workers/Email/Validated.cs
@@ -1,5 +1,5 @@
 ﻿// -----------------------------------------------------------------------
-//  <copyright file = "EmailValidated.cs" company = "Prism">
+//  <copyright file = "Validated.cs" company = "Prism">
 //  Copyright (c) Prism.All rights reserved.
 //  </copyright>
 // -----------------------------------------------------------------------
@@ -10,18 +10,18 @@ using Microsoft.Azure.Functions.Worker;
 using Prism.Picshare.Commands.Authentication;
 using Prism.Picshare.Events;
 
-namespace Prism.Picshare.AzureServices.Workers.User;
+namespace Prism.Picshare.AzureServices.Workers.Email;
 
-public class EmailValidated
+public class Validated
 {
     private readonly IMediator _mediator;
 
-    public EmailValidated(IMediator mediator)
+    public Validated(IMediator mediator)
     {
         _mediator = mediator;
     }
 
-    [Function(nameof(User) + "." + nameof(EmailValidated))]
+    [Function(nameof(Email) + "." + nameof(Validated))]
     public async Task<ResultCodes> Run([ServiceBusTrigger(Topics.Email.Validated, Connection = "SERVICE_BUS_CONNECTION_STRING")] string mySbMsg, FunctionContext context)
     {
         var user = JsonSerializer.Deserialize<Domain.User>(mySbMsg);
