+++ src/net/services/Prism.Picshare.AzureServices.Workers/Pictures/Created.cs
@@ -7,24 +7,22 @@
 using System.Text.Json;
 using MediatR;
 using Microsoft.Azure.Functions.Worker;
-using Prism.Picshare.AzureServices.Workers.Mailing;
-using Prism.Picshare.Commands.Mailing;
 using Prism.Picshare.Commands.Processor;
 using Prism.Picshare.Domain;
 using Prism.Picshare.Events;
 
-namespace Prism.Picshare.AzureServices.Workers.Processor;
+namespace Prism.Picshare.AzureServices.Workers.Pictures;
 
-public class PictureCreated : ISimpleFunction
+public class Created : ISimpleFunction
 {
     private readonly IMediator _mediator;
 
-    public PictureCreated(IMediator mediator)
+    public Created(IMediator mediator)
     {
         _mediator = mediator;
     }
 
-    [Function(nameof(Processor) + "." + nameof(PictureCreated))]
+    [Function(nameof(Pictures) + "." + nameof(Created))]
     public async Task Run([ServiceBusTrigger(Topics.Pictures.Created, Connection = "SERVICE_BUS_CONNECTION_STRING")] string mySbMsg, FunctionContext context)
     {
         var picture = JsonSerializer.Deserialize<EntityReference>(mySbMsg);
