+++ src/net/services/Prism.Picshare.AzureServices.Workers/User/Register.cs
@@ -10,18 +10,18 @@ using Microsoft.Azure.Functions.Worker;
 using Prism.Picshare.Commands.Mailing;
 using Prism.Picshare.Events;
 
-namespace Prism.Picshare.AzureServices.Workers.Mailing;
+namespace Prism.Picshare.AzureServices.Workers.User;
 
-public class UserRegister: ISimpleFunction
+public class Register: ISimpleFunction
 {
     private readonly IMediator _mediator;
 
-    public UserRegister(IMediator mediator)
+    public Register(IMediator mediator)
     {
         _mediator = mediator;
     }
 
-    [Function(nameof(Mailing) + "." + nameof(UserRegister))]
+    [Function(nameof(User) + "." + nameof(Register))]
     public async Task Run([ServiceBusTrigger(Topics.User.Register, Connection = "SERVICE_BUS_CONNECTION_STRING")] string mySbMsg, FunctionContext context)
     {
         var user = JsonSerializer.Deserialize<Domain.User>(mySbMsg);
